Params and GUI both support sample rate multiplier

+++ include/vapor/RayCasterParams.h
@@ -35,10 +35,10 @@ public:
     // Different ray casting methods: 1 == fixed step casting
     //                                2 == prism intersection casting
     //
-    long   GetCastingMode() const;
-    void   SetCastingMode(long);
-    double GetSampleMultiplier() const;
-    void   SetSampleRateMultiplier(double);
+    long GetCastingMode() const;
+    void SetCastingMode(long);
+    long GetSampleMultiplier() const;      // ComboBox index is held here. Need to translate
+    void SetSampleRateMultiplier(long);    //   to real multipliers in RayCaster.cpp
 
 protected:
     static const std::string _lightingTag;
